Fix Win64 compilation failure introduced by r10483.
R=erik.corry@gmail.com

Review URL: https://chromiumcodereview.appspot.com/9284017

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10485 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/deoptimizer.cc
@@ -632,7 +632,7 @@ void Deoptimizer::MaterializeHeapNumbersForDebuggerInspectableFrame(
       Handle<Object> num = isolate_->factory()->NewNumber(d.value());
 
       int index = (info->parameters_count() - 1) -
-          (slot - parameters_top) / kPointerSize;
+          static_cast<int>(slot - parameters_top) / kPointerSize;
 
       if (FLAG_trace_deopt) {
         PrintF("Materializing a new heap number %p [%e] in slot %p"
@@ -648,7 +648,7 @@ void Deoptimizer::MaterializeHeapNumbersForDebuggerInspectableFrame(
       Handle<Object> num = isolate_->factory()->NewNumber(d.value());
 
       int index = info->expression_count() - 1 -
-          (slot - expressions_top) / kPointerSize;
+          static_cast<int>(slot - expressions_top) / kPointerSize;
 
       if (FLAG_trace_deopt) {
         PrintF("Materializing a new heap number %p [%e] in slot %p"
